task5/back/requirements.txt
mysql-connector-python
jinja2
bcrypt
bson
